Added Books section + one title
Hi,

Stjepan from Manning here. I thought this title might be a good match for your list, so I've added a new category for the Books and added the title. Thank you for considering it. If you're interested in having more books on similar subjects, please, let me know because Mannig has quite a few titles on the subject.

Best,